Re-org files a bit; update module docstring
